requirements.txt
flask
flask_cors
openai
python-dotenv
markdown